% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaixaDados.R
\name{BaixaDados}
\alias{BaixaDados}
\title{Baixa dados de matrix do economatica}
\usage{
BaixaDados(Nome, PathFile, Periodo, Planilha)
}
\arguments{
\item{Nome}{Nome da variavel}

\item{PathFile}{Caminho do aquivo}
}
\value{
o valor...
}
\description{
Mais detalhes
}
\examples{

}
