% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contabilidados.R
\docType{package}
\name{contabilidados}
\alias{contabilidados}
\alias{contabilidados-package}
\title{Pacote CONTABILIDADOS}
\description{
Pacote voltado a disseminacao de funcoes criadas pela equipe CONTABILIDADOS com
o intuito de facilitar coleta, tratamento e estimacao de dados para pesquisas
academicas em contabilidade, economia e administracao.
}
\author{
Kleber Formiga Miranda \email{mirandakf@ufersa.edu.br}
}
