% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contabilidados.R
\docType{package}
\name{contabilidados}
\alias{contabilidados}
\alias{contabilidados-package}
\title{Pacote CONTABILIDADOS}
\description{
Fazer a descricao da funcao
completando aqui em baixo
}
\author{
Kleber Formiga Miranda \email{mirandakf@ufersa.edu.br}
}
